Make them not be variables
